+++ hw02-annotation-based/src/main/java/ru/otus/homework/quiz/service/QuizTestService.java
@@ -1,0 +1,9 @@
+package ru.otus.homework.quiz.service;
+
+import java.util.List;
+import ru.otus.homework.quiz.domain.QuizQuestion;
+
+public interface QuizTestService {
+
+  void runTest(List<QuizQuestion> quizQuestions);
+}
